Crate quilkin

Re-exports

pub use providers::Providers;

pub use service::Service;

pub use quilkin_proto as generated;

Modules

alloc

cli

codec

Implementations and utility methods for various codecs used in Quilkin.

collections

Collection types designed for use with Quilkin.

components

config

Quilkin configuration.

filters

Filters for processing packets.

metrics

net

providers

service

signal

time

xds

Macros

__func_name

Macro that can get the function name of the function the macro is invoked within

include_proto

Includes generated Protobuf definitions from tonic.

temp_file

Creates a temporary file with the specified prefix in a directory named after the calling function, ie using it within a test will place it in a temporary directory named after the test

Structs

Cli

Quilkin: a non-transparent UDP proxy specifically designed for use with large scale multiplayer dedicated game servers deployments, to ensure security, access control, telemetry data, metrics and more.

Config

Type Aliases

Result