quilkin
Module cli

Re-exports

pub use self::agent::Agent;

pub use self::generate_config_schema::GenerateConfigSchema;

pub use self::manage::Manage;

pub use self::proxy::Proxy;

pub use self::qcmp::Qcmp;

pub use self::relay::Relay;

Modules

agent

generate_config_schema

manage

proxy

qcmp

relay

Structs

AdminCli

Cli

Quilkin: a non-transparent UDP proxy specifically designed for use with large scale multiplayer dedicated game servers deployments, to ensure security, access control, telemetry data, metrics and more.

LocalityCli

Service

Enums

Commands

The various Quilkin commands.

LogFormats

The various log format options