quilkin::cli
Struct Service

pub struct Service {
    pub xdp: XdpOptions,
    pub udp_workers: NonZeroUsize,
    /* private fields */
}

Fields
xdp: XdpOptionsudp_workers: NonZeroUsize

Amount of UDP workers to run.

Implementations

impl Service

pub fn udp(self) -> Self

Enables the UDP service.

pub fn udp_port(self, port: u16) -> Self

Sets the UDP service port.

pub fn qcmp(self) -> Self

Enables the QCMP service.

pub fn qcmp_port(self, port: u16) -> Self

Sets the UDP service port.

pub fn mds(self) -> Self

Enables the mDS service.

pub fn mds_port(self, port: u16) -> Self

Sets the mDS service port.

pub fn phoenix(self) -> Self

Enables the Phoenix service.

pub fn phoenix_port(self, port: u16) -> Self

Sets the Phoenix service port.

pub fn xds(self) -> Self

Enables the xDS service.

pub fn xds_port(self, port: u16) -> Self

Sets the xDS service port.

pub fn spawn_services( self, config: &Arc<Config>, shutdown_rx: &ShutdownRx, ) -> Result<JoinHandle<Result<()>>>

The main entrypoint for listening network servers. When called will spawn any and all enabled services, if successful returning a future that can be await to wait on services to be cancelled.

pub fn publish_udp( &self, config: &Arc<Config>, ) -> Result<(impl Future<Output = Result<()>>, Option<Box<dyn FnOnce(ShutdownRx) + Send>>)>

pub fn spawn_user_space_router( &self, config: Arc<Config>, ) -> Result<(impl Future<Output = Result<()>>, Box<dyn FnOnce(ShutdownRx) + Send>)>

Launches the user space implementation of the packet router using sockets. This implementation uses a pool of buffers and sockets to manage UDP sessions and sockets. On Linux this will use io-uring, where as it will use epoll interfaces on non-Linux platforms.