quilkin::codec

Module qcmp

Logic for parsing and generating Quilkin Control Message Protocol (QCMP) messages.

Structs

QcmpPacket

QcmpTransceiver

A transciever that can handle multiple simultaneous QCMP pings over the same socket and ensure that responses are forwarded to the correct receiver via the QCMP nonce

Enums

Error

Protocol

The set of possible QCMP commands.

Constants

MAX_QCMP_PACKET_LEN

The maximum length of a QCMP packet, including 2 additional i64 timestamps

MIN_QCMP_PACKET_LEN

The minimum length of a QCMP packet

Functions

port_channel

spawn