Module quilkin::components::proxy

Modules

packet_router

Structs

PendingSends

A simple packet queue that signals when a packet is pushed

Proxy

Ready

RecvPacket

SendPacket

SessionPool

A data structure that is responsible for holding sessions, and pooling sockets between them. This means that we only provide new unique sockets to new connections to the same gameserver, and we share sockets across multiple gameservers.

ToTokens

Enums

PipelineError

Type Aliases

ErrorMap