Module quilkin::config

Quilkin configuration.

Re-exports

pub use self::providers::Providers;

pub use self::watch::Watch;

Modules

providers

watch

Structs

Config

Config is the configuration of a proxy

DeltaDiscoveryRes

Filter

Filter is the configuration for a single filter

Slot

A mutable memory location with atomic storage rules.

Enums

Base64Standard

ConfigType

The configuration of a Filter from either a static or dynamic source.

ValidationError

Validation failure for a Config

Version

Functions

max_grpc_message_size

Returns the configured maximum allowed message size for gRPC messages. When using State Of The World xDS, the message size can get large enough that it can exceed the default limits.