quilkin::config::providers::k8s
Module agones

Structs

Fleet

FleetInner

FleetSpec

FleetSpec is the spec for a Fleet. More info: https://agones.dev/site/docs/reference/agones_crd_api_reference/#agones.dev/v1.Fleet

FleetStatus

FleetStatus is the status of a Fleet. More info: https://agones.dev/site/docs/reference/agones_crd_api_reference/#agones.dev/v1.Fleet

FleetStrategy

FleetStrategyRollingUpdate

GameServer

Auto-generated derived type for GameServerSpec via CustomResource

GameServerPort

Defines a set of Ports that are to be exposed via the GameServer.

GameServerSpec

GameServerStatus

The status for a GameServer resource.

GameServerStatusPort

The port that was allocated to a GameServer.

GameServerTemplateSpec

GameServer is the data structure for a GameServer resource.

Health

Inner

SdkServer

Parameters for the Agones SDK Server sidecar container

Enums

FleetScheduling

FleetSpec is the spec for a Fleet. More info: https://agones.dev/site/docs/reference/agones_crd_api_reference/#agones.dev/v1.Fleet

FleetStrategyType

GameServerState

PortPolicy

Protocol

SchedulingStrategy

the strategy that a Fleet & GameServers will use when scheduling GameServers’ Pods across a cluster. In future versions, this will also impact Fleet scale down, and Pod Scheduling.

SdkServerLogLevel