Struct quilkin::config::Slot

pub struct Slot<T> { /* private fields */ }

A mutable memory location with atomic storage rules.

Implementations

impl<T> Slot<T>

pub fn new(value: impl Into<Option<T>>) -> Self

Creates a new slot for value.

pub fn empty() -> Self

Creates a new empty slot.

pub fn watch(&self, watcher: impl Fn(&T) + Send + Sync + 'static)

Adds a watcher to to the slot. The watcher will fire whenever slot’s value changes.

pub fn is_some(&self) -> bool

Returns whether any data is present in the slot.

impl<T: Default> Slot<T>

pub fn with_default() -> Self

Creates a new slot containing the default instance of T.

pub fn load(&self) -> Arc<T>

Provides a reference to the underlying data.

pub fn store(&self, value: Arc<T>)

Replaces the data in the slot with value.

pub fn remove(&self)

Removes any data from the slot.

pub fn store_if_unset(&self, value: Arc<T>)

Replaces the data if the slot is empty.

impl<T: Default + PartialEq> Slot<T>

pub fn try_replace(&self, value: Self)

Replaces the current data in the slot with value’s data, if present.

impl<T: Clone + Default> Slot<T>

pub fn modify(&self, modify: impl FnMut(&mut T))

Provides a view into a mutable reference of the current data in the slot. Any changes made will update the value in the slot.