Module quilkin::config::watch::agones::crd

Structs

GameServer

Auto-generated derived type for GameServerSpec via CustomResource

GameServerPort

Defines a set of Ports that are to be exposed via the GameServer.

GameServerSpec

GameServerStatus

The status for a GameServer resource.

GameServerStatusPort

The port that was allocated to a GameServer.

Health

Inner

SdkServer

Parameters for the Agones SDK Server sidecar container

Enums

GameServerState

PortPolicy

Protocol

SchedulingStrategy

the strategy that a Fleet & GameServers will use when scheduling GameServers’ Pods across a cluster. In future versions, this will also impact Fleet scale down, and Pod Scheduling.

SdkServerLogLevel