Module quilkin::endpoint

Types representing where the data is the sent.

Structs

Endpoint

A destination endpoint with any associated metadata.

EndpointAddress

A valid socket address. This differs from std::net::SocketAddr, in that it it supports parsing Domain Names in addition to IP addresses. Domain Names are resolved when ToSocketAddrs::to_socket_addrs is called.

Locality

The location of an Endpoint.

LocalityEndpoints

A set of endpoints optionally grouped by a Locality.

LocalitySet

Set around LocalityEndpoints to ensure that all unique localities are different entries. Any duplicate localities provided are merged.

Metadata

Metadata specific to endpoints.

Enums

MetadataError

Type Definitions

EndpointMetadata