Struct quilkin::endpoint::LocalityEndpoints

pub struct LocalityEndpoints {
    pub locality: Option<Locality>,
    pub endpoints: BTreeSet<Endpoint>,
}

A set of endpoints optionally grouped by a Locality.

Fields
locality: Option<Locality>endpoints: BTreeSet<Endpoint>
Implementations

impl LocalityEndpoints

pub fn new(endpoints: BTreeSet<Endpoint>) -> Self

Creates a new set of endpoints with no Locality.

pub fn with_locality(self, locality: impl Into<Option<Locality>>) -> Self

Adds a Locality to the set of endpoints.

pub fn remove(&mut self, endpoint: &Endpoint)

Removes an endpoint.