Struct quilkin::endpoint::LocalitySet

pub struct LocalitySet(_);

Set around LocalityEndpoints to ensure that all unique localities are different entries. Any duplicate localities provided are merged.

Implementations

impl LocalitySet

pub fn new(set: Vec<LocalityEndpoints>) -> Self

Creates a new set from the provided localities.

pub fn insert(&mut self, locality: LocalityEndpoints)

Inserts a new locality of endpoints.

pub fn get(&self, key: &Option<Locality>) -> Option<&LocalityEndpoints>

Returns a reference to endpoints associated with the specified locality.

pub fn get_mut( &mut self, key: &Option<Locality> ) -> Option<&mut LocalityEndpoints>

Returns a mutable reference to endpoints associated with the specified locality.

pub fn remove(&mut self, key: &Option<Locality>) -> Option<LocalityEndpoints>

Removes the specified locality or all endpoints with no locality.

pub fn clear(&mut self)

Removes all localities.

pub fn iter(&self) -> impl Iterator<Item = &LocalityEndpoints> + '_

Returns an iterator over the set of localities.

pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut LocalityEndpoints> + '_

Returns a mutable iterator over the set of localities.

pub fn merge(&mut self, cluster: &Self)