quilkin::filters
Module capture

Structs

Capture

Config

Prefix

Capture from the start of the packet.

Regex

Capture from the start of the packet.

Suffix

Capture from the end of the packet.

Enums

Strategy

Strategy to apply for acquiring a set of bytes in the UDP packet

Constants

CAPTURED_BYTES

The default key under which the Capture filter puts the byte slices it extracts from each packet.

Traits

CaptureStrategy

Trait to implement different strategies for capturing packet data.