quilkin::filters

Module firewall

Structs

Cidr

Cidr notation for an ipv6 or ipv4 netmask

Config

Represents how a Firewall filter is configured for read and write operations.

Firewall

Filter for allowing/blocking traffic by IP and port.

PortRange

Range of matching ports that are configured against a Rule.

Rule

Combination of CIDR range, port range and action to take.

Enums

Action

Whether or not a matching Rule should Allow or Deny access

IpNetwork

Represents a generic network range. This type can have two variants: the v4 and the v6 case.

PortRangeError

Invalid min and max values for a PortRange.