Module quilkin::filters::prelude

Prelude containing all types and traits required to implement Filter and FilterFactory.

Re-exports

pub use super::Filter;

Structs

ConvertProtoConfigError
An error representing failure to convert a filter’s protobuf configuration to its static representation.

CreateFilterArgs
Arguments needed to create a new filter.

ReadContext
The input arguments to Filter::read.

ReadResponse
The output of Filter::read.

WriteContext
The input arguments to Filter::write.

WriteResponse
The output of Filter::write.

Enums

Error
An error that occurred when attempting to create a Filter from a FilterFactory.

Traits

FilterFactory
Provides the name and creation function for a given Filter.

Type Definitions

DynFilterFactory
An owned pointer to a dynamic FilterFactory instance.