Trait quilkin::filters::prelude::FilterFactory

pub trait FilterFactory: Sync + Send {
    fn name(&self) -> &'static str;

    fn create_filter(
        &self, 
        args: CreateFilterArgs<'_>
    ) -> Result<Box<dyn Filter>, Error>;

    fn require_config<'a, 'b>(
        &'a self, 
        config: Option<ConfigType<'b>>
    ) -> Result<ConfigType<'b>, Error> { ... }
}

Provides the name and creation function for a given Filter.

Required methods

fn name(&self) -> &'static str

name returns the configuration name for the Filter The returned string identifies the filter item’s path with the following format: quilkin.extensions.filters.<module>.<version>.<item-name> where: : The rust module name containing the filter item : The filter’s version. : The name of the rust item (e.g enum, struct) implementing the filter. For example the v1alpha1 version of the debug filter has the name: quilkin.extensions.filters.debug_filter.v1alpha1.Debug

fn create_filter(
&self,
args: CreateFilterArgs<'_>
) -> Result<Box<dyn Filter>, Error>

Returns a filter based on the provided arguments.

Provided methods

fn require_config<'a, 'b>(
&'a self,
config: Option<ConfigType<'b>>
) -> Result<ConfigType<'b>, Error>

Returns the ConfigType from the provided Option, otherwise it returns Error::MissingConfig if the Option is None.