Struct quilkin::filters::FilterChain

pub struct FilterChain { /* private fields */ }

A chain of Filters to be executed in order.

Executes each filter, passing the ReadContext and WriteContext between each filter’s execution, returning the result of data that has gone through all of the filters in the chain. If any of the filters in the chain return None, then the chain is broken, and None is returned.

Implementations

impl FilterChain

pub fn new( filters: Vec<(String, FilterInstance)> ) -> Result<Self, CreationError>

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn iter(&self) -> impl Iterator<Item = Filter> + '_

pub fn try_create_fallible<Item>( filter_configs: impl IntoIterator<Item = Item> ) -> Result<Self, CreationError>
where Item: TryInto<FilterConfig, Error = CreationError>,

Validates the filter configurations in the provided config and constructs a FilterChain if all configurations are valid, including the conversion into a Filter

pub fn try_create( filter_configs: impl IntoIterator<Item = FilterConfig> ) -> Result<Self, CreationError>

Validates the filter configurations in the provided config and constructs a FilterChain if all configurations are valid.