Struct quilkin::filters::FilterSet

pub struct FilterSet(_);

A set of filters to be registered with a FilterRegistry.

Implementations

impl FilterSet

pub fn default() -> Self

Returns a default set of filters which are runtime configurable and used with each endpoint.

Current default filters:

debug

local_rate_limit

concatenate_bytes

load_balancer

capture_bytes

token_router

compress

pub fn default_with(filters: impl IntoIterator<Item = DynFilterFactory>) -> Self

Returns a FilterSet with the filters provided through filters in addition to the defaults. Any filter factories provided by filters will override any defaults with a matching name.

See FilterSet::default for a list of the current defaults.

pub fn with(filters: impl IntoIterator<Item = DynFilterFactory>) -> Self

Creates a new FilterSet with the set of filter_factories without any defaults.