Struct quilkin::filters::LocalRateLimit

pub struct LocalRateLimit { /* private fields */ }

A filter that implements rate limiting on packets based on the token-bucket algorithm. Packets that violate the rate limit are dropped. It only applies rate limiting on packets received from a downstream connection (processed through LocalRateLimit::read). Packets coming from upstream endpoints flow through the filter untouched.