Struct quilkin::filters::timestamp::Timestamp

pub struct Timestamp { /* private fields */ }

A filter that reads a metadata value as a timestamp to be observed in a histogram.

Implementations

impl Timestamp

pub fn observe(&self, metadata: &DynamicMetadata, direction: Direction)

Observes the duration since a timestamp stored in metadata and now, if present.