quilkin::filters
Trait StaticFilter

pub trait StaticFilter:
    Filter
    + Sized
    + Into<FilterKind>
where
    CreationError: From<<Self::Configuration as TryFrom<Self::BinaryConfiguration>>::Error> + From<<Self::BinaryConfiguration as TryFrom<Self::Configuration>>::Error>,
{
    type Configuration: JsonSchema + Serialize + for<'de> Deserialize<'de> + TryFrom<Self::BinaryConfiguration>;
    type BinaryConfiguration: Message + Default + TryFrom<Self::Configuration> + Send + Sync + Sized;

    const NAME: &'static str;

    // Required method
    fn try_from_config(
        config: Option<Self::Configuration>,
    ) -> Result<Self, CreationError>;

    // Provided methods
    fn from_config(config: Option<Self::Configuration>) -> Self { ... }
    fn factory() -> DynFilterFactory
       where Self: 'static { ... }
    fn ensure_config_exists(
        config: Option<Self::Configuration>,
    ) -> Result<Self::Configuration, CreationError> { ... }
    fn as_filter_config(
        config: impl Into<Option<Self::Configuration>>,
    ) -> Result<Filter, CreationError> { ... }
    fn as_labeled_filter_config(
        config: impl Into<Option<Self::Configuration>>,
        label: String,
    ) -> Result<Filter, CreationError> { ... }
}

Statically safe version of Filter, if you’re writing a Rust filter, you should implement StaticFilter in addition to Filter, as StaticFilter guarantees all of the required properties through the type system, allowing Quilkin take care of the virtual table boilerplate automatically at compile-time.

Required Associated Constants

const NAME: &'static str

The globally unique name of the filter.

Required Associated Types

type Configuration: JsonSchema + Serialize + for<'de> Deserialize<'de> + TryFrom<Self::BinaryConfiguration>

The human-readable configuration of the filter. Must be serde compatible, have a JSON schema, and be convertible to and from Self::BinaryConfiguration.

type BinaryConfiguration: Message + Default + TryFrom<Self::Configuration> + Send + Sync + Sized

The binary configuration of the filter. Must be prost compatible, and be convertible to and from Self::Configuration.

Required Methods

fn try_from_config( config: Option<Self::Configuration>, ) -> Result<Self, CreationError>

Instantiates a new StaticFilter from the given configuration, if any.

Errors

If the provided configuration is invalid.

Provided Methods

fn from_config(config: Option<Self::Configuration>) -> Self

Instantiates a new StaticFilter from the given configuration, if any.

Panics

If the provided configuration is invalid.

fn factory() -> DynFilterFactory
where Self: 'static,

Creates a new dynamic FilterFactory virtual table.

fn ensure_config_exists( config: Option<Self::Configuration>, ) -> Result<Self::Configuration, CreationError>

Convenience method for providing a consistent error message for filters which require a fully initialized Self::Configuration.

fn as_filter_config( config: impl Into<Option<Self::Configuration>>, ) -> Result<Filter, CreationError>

fn as_labeled_filter_config( config: impl Into<Option<Self::Configuration>>, label: String, ) -> Result<Filter, CreationError>

Object Safety

This trait is not object safe.