Enum quilkin::metadata::Symbol

pub enum Symbol {
    Reference(Reference),
    Literal(Value),
}

A literal value or a reference to a value in a metadata map.

Variants

Reference(Reference)

Literal(Value)

Implementations

impl Symbol

pub fn as_literal(&self) -> Option<&Value>

pub fn as_reference(&self) -> Option<&Reference>

pub fn resolve<'literal: 'metadata, 'metadata>( &'literal self, metadata: &'metadata DynamicMetadata ) -> Option<&'metadata Value>

Resolves a symbol into a Value, using ctx for any references, returning None if could not be found.

pub fn resolve_to_bytes<'literal: 'metadata, 'metadata>( &'literal self, metadata: &'metadata DynamicMetadata ) -> Option<Bytes>

Tries to Self::resolve the symbol to a bytes::Bytes, performing a conversion process on different Value if relevant. Returning None if it isn’t supported currently.

Value::Bytes The value is copied as-is.

Value::String The value is interpreted as a base64 string.

Value::Number The value is an eight byte number encoded as big endian.