quilkin

Module metrics

Structs

AsnInfo

Enums

Direction

Constants

DIRECTION_LABEL

“event” is used as a label for Metrics that can apply to both Filter read and write executions.

EMPTY

READ_DIRECTION_LABEL

Label value for DIRECTION_LABEL for read events

WRITE_DIRECTION_LABEL

Label value for DIRECTION_LABEL for write events

Traits

CollectorExt

Functions

histogram_opts

opts

Create a generic metrics options. Use filter_opts instead if the intended target is a filter.

register

Registers the current metric collector with the provided registry.

register_metrics

registry

Returns the [Registry] containing all the metrics registered in Quilkin.

with_mut_registry

with_registry

Type Aliases

Result

A specialized Result type for prometheus.