Module quilkin::net

Re-exports

pub use self::cluster::ClusterMap;

pub use self::endpoint::Endpoint;

pub use self::endpoint::EndpointAddress;

Modules

cluster

endpoint

Types representing where the data is the sent.

Structs

DualStackLocalSocket

An ipv6 socket that can accept and send data from either a local ipv4 address or ipv6 address with port reuse enabled and only_v6 set to false.

Functions

to_canonical

Converts a a socket address to its canonical version. This is just a copy of the method available in std but that is currently nightly only.