quilkin::net::io::nic::xdp

Function setup_xdp_io

pub fn setup_xdp_io(config: XdpConfig<'_>) -> Result<XdpWorkers, XdpSetupError>

a socket is bound to every available queue on the NIC, and when spawn is invoked, each socket is processed in its own thread

Binding to fewer queues is possible in the future but requires additional work in the xdp crate