quilkin::net

Module packet

Re-exports

pub use self::queue::PacketQueue;

pub use self::queue::PacketQueueSender;

pub use self::queue::queue;

Modules

queue

Traits

Packet

Representation of an immutable set of bytes pulled from the network, this trait provides an abstraction over however the packet was received (epoll, io-uring, xdp)

PacketMut

Representation of an mutable set of bytes pulled from the network, this trait provides an abstraction over however the packet was received (epoll, io-uring, xdp)

Functions

spawn_receivers

Spawns a background task that sits in a loop, receiving packets from the passed in socket. Each received packet is placed on a queue to be processed by a worker task. This function also spawns the set of worker tasks responsible for consuming packets off the aforementioned queue and processing them through the filter chain and session pipeline.