quilkin::net::packet

Trait Packet

pub trait Packet: Sized {
    // Required methods
    fn as_slice(&self) -> &[u8] ;
    fn len(&self) -> usize;

    // Provided method
    fn is_empty(&self) -> bool { ... }
}

Representation of an immutable set of bytes pulled from the network, this trait provides an abstraction over however the packet was received (epoll, io-uring, xdp)

Use PacketMut if you need a mutable representation.

Required Methods

fn as_slice(&self) -> &[u8]

Returns the underlying slice of bytes representing the packet.

fn len(&self) -> usize

Returns the size of the packet.

Provided Methods

fn is_empty(&self) -> bool

Returns whether the given packet is empty.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.