quilkin::net::packet

Trait PacketMut

pub trait PacketMut: Sized + Packet {
    // Required methods
    fn remove_head(&mut self, length: usize);
    fn remove_tail(&mut self, length: usize);
    fn extend_head(&mut self, bytes: &[u8]);
    fn extend_tail(&mut self, bytes: &[u8]);
    fn freeze(self) -> Bytes;
}

Representation of an mutable set of bytes pulled from the network, this trait provides an abstraction over however the packet was received (epoll, io-uring, xdp)

Required Methods

fn remove_head(&mut self, length: usize)

Truncates the head by the specified number of bytes

fn remove_tail(&mut self, length: usize)

Truncates the tail by the specified number of bytes

fn extend_head(&mut self, bytes: &[u8])

Prepend the specified bytes

fn extend_tail(&mut self, bytes: &[u8])

Append the specified bytes

fn freeze(self) -> Bytes

Returns an immutable version of the packet, this allows certain types return a type that can be more cheaply cloned and shared.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.