Struct quilkin::net::phoenix::Phoenix

pub struct Phoenix<M> { /* private fields */ }

A Phoenix instance maintains a virtual coordinate space for nodes in a distributed system to estimate their network latencies. It uses the provided Measurement trait to periodically measure and update each node’s coordinates, allowing for latency estimation between any two nodes.

Implementations

impl<M: Measurement + 'static> Phoenix<M>

pub fn new(measurement: M) -> Self

pub fn builder(measurement: M) -> Builder<M>

pub async fn background_update_task(&self)

Starts the background update task to continously sample from nodes and update their coordinates.

pub fn get_coordinates(&self, address: &SocketAddr) -> Option<Coordinates>

pub fn ordered_nodes_by_latency(&self) -> Vec<(IcaoCode, f64)>

pub fn add_node(&self, address: SocketAddr, icao_code: IcaoCode)

pub fn add_node_if_not_exists(&self, address: SocketAddr, icao_code: IcaoCode)

pub fn add_nodes_from_config(&self, config: &Config)