quilkin::net::phoenix
Trait Measurement

pub trait Measurement {
    // Required method
    fn measure_distance<'life0, 'async_trait>(
        &'life0 self,
        address: SocketAddr,
    ) -> Pin<Box<dyn Future<Output = Result<DistanceMeasure>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

An implementation of measuring the network difference between two nodes.

Required Methods

fn measure_distance<'life0, 'async_trait>( &'life0 self, address: SocketAddr, ) -> Pin<Box<dyn Future<Output = Result<DistanceMeasure>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets the difference between this node and address, returning the latency in nanoseconds on success.