quilkin::net::sessions

Trait SessionManager

pub trait SessionManager {
    // Required method
    fn send(
        &self,
        key: SessionKey,
        contents: Bytes,
    ) -> Result<(), PipelineError>;
}

Responsible for managing sending processed traffic to its destination and tracking metrics and other information about the session.

Required Methods

fn send(&self, key: SessionKey, contents: Bytes) -> Result<(), PipelineError>