Module quilkin::net::xds

Modules

core

discovery

listener

socket_address

Nested message and enum types in SocketAddress.

Structs

Client

Client that can talk to an XDS server using the aDS protocol.

Enums

AwaitingAck

The resources and versions that were sent in a delta response, when acked this is used with ClientVersions::ack to update the set of resources on the client has

ClientVersions

Keeps track of what resource versions a particular client has

Resource

ResourceType

Type Aliases

AdsClient