Enum quilkin::protocol::Protocol

pub enum Protocol {
    Ping {
        client_timestamp: i64,
        nonce: u8,
    },
    PingReply {
        client_timestamp: i64,
        nonce: u8,
        server_start_timestamp: i64,
        server_transmit_timestamp: i64,
    },
}

The set of possible QCMP commands.

Variants

Ping

Fields

client_timestamp: i64

The timestamp from when the client sent the packet.

nonce: u8

The client’s nonce.

The initation of a ping command to send to a Quilkin proxy to measure latency.

PingReply

Fields

client_timestamp: i64

The timestamp from when the client sent the ping packet.

nonce: u8

The client’s nonce.

server_start_timestamp: i64

The timestamp from when the server received the ping packet.

server_transmit_timestamp: i64

The timestamp from when the server sent the reply.

The reply from a Quilkin proxy from a Self::Ping command. Can be used with Protocol::round_trip_delay to measure latency between two machines.

Implementations

impl Protocol

pub fn ping() -> Self

Creates a Self::Ping with a random nonce, should be sent as soon as possible from creation to maintain accuracy.

pub fn ping_with_nonce(nonce: u8) -> Self

Creates a Self::Ping with a user-specified nonce, should be sent as soon as possible from creation to maintain accuracy.

pub fn ping_reply(
nonce: u8,
client_timestamp: i64,
server_start_timestamp: i64
) -> Self

Creates a Self::PingReply from the client and server start timestamp. It’s recommended to transmit as as soon as possible from creation to keep the start and transmit times as accurate as possible.

pub fn encode(&self) -> Vec<u8>

Encodes the protocol command into a buffer of bytes for network transmission.

pub fn nonce(&self) -> u8

Returns the packet’s nonce.

pub fn round_trip_delay(&self, client_response_timestamp: i64) -> Option<i64>

If the command is Protocol::PingReply, with client_response_timestamp measures the roundtrip delay between the client machine, and the Quilkin proxy, using the same algorithm as Network Time Protocol (NTP).

pub fn parse(input: &[u8]) -> Result<Option<Self>, Error>

Parses the provided input, and attempts to parse it as a Protocol packet. Returning None if the magic number is not present, and thus is not a QCMP packet, and returning Err when it was detected as a QCMP packet, but there was an error in parsing the payload.