Struct quilkin::Proxy

pub struct Proxy {
    pub management_server: Vec<Endpoint>,
    pub mmdb: Option<Source>,
    pub port: u16,
    pub to: Vec<SocketAddr>,
}

Run Quilkin as a UDP reverse proxy.

Fields
management_server: Vec<Endpoint>

One or more quilkin manage endpoints to listen to for config changes

mmdb: Option<Source>

The remote URL or local file path to retrieve the Maxmind database.

port: u16

The port to listen on.

to: Vec<SocketAddr>

One or more socket addresses to forward packets to.

Implementations

impl Proxy

pub async fn run(
&self,
config: Arc<Config>,
shutdown_rx: Receiver<()>
) -> Result<()>

Start and run a proxy.