Module quilkin::xds

Modules

config

core

service

type

Structs

AggregatedDiscoveryServiceClient

See https://github.com/envoyproxy/envoy-api#apis for a description of the role of ADS and how it is intended to be used by a management server. ADS requests have the same structure as their singleton xDS counterparts, but can multiplex many resource types on a single stream. The type_url in the DiscoveryRequest/DiscoveryResponse provides sufficient information to recover the multiplexed singleton APIs at the Envoy instance and management server.

Client

Client that can talk to an XDS server using the aDS protocol.

ControlPlane

Enums

Resource

ResourceType