Struct quilkin::xds::config::cluster::v3::Cluster

pub struct Cluster {

    pub transport_socket_matches: Vec<TransportSocketMatch>,
    pub name: String,
    pub alt_stat_name: String,
    pub eds_cluster_config: Option<EdsClusterConfig>,
    pub connect_timeout: Option<Duration>,
    pub per_connection_buffer_limit_bytes: Option<u32>,
    pub lb_policy: i32,
    pub load_assignment: Option<ClusterLoadAssignment>,
    pub health_checks: Vec<HealthCheck>,
    pub max_requests_per_connection: Option<u32>,
    pub circuit_breakers: Option<CircuitBreakers>,
    pub upstream_http_protocol_options: Option<UpstreamHttpProtocolOptions>,
    pub common_http_protocol_options: Option<HttpProtocolOptions>,
    pub http_protocol_options: Option<Http1ProtocolOptions>,
    pub http2_protocol_options: Option<Http2ProtocolOptions>,
    pub typed_extension_protocol_options: HashMap<String, Any>,
    pub dns_refresh_rate: Option<Duration>,
    pub dns_failure_refresh_rate: Option<RefreshRate>,
    pub respect_dns_ttl: bool,
    pub dns_lookup_family: i32,
    pub dns_resolvers: Vec<Address>,
    pub use_tcp_for_dns_lookups: bool,
    pub dns_resolution_config: Option<DnsResolutionConfig>,
    pub typed_dns_resolver_config: Option<TypedExtensionConfig>,
    pub wait_for_warm_on_init: Option<bool>,
    pub outlier_detection: Option<OutlierDetection>,
    pub cleanup_interval: Option<Duration>,
    pub upstream_bind_config: Option<BindConfig>,
    pub lb_subset_config: Option<LbSubsetConfig>,
    pub common_lb_config: Option<CommonLbConfig>,
    pub transport_socket: Option<TransportSocket>,
    pub metadata: Option<Metadata>,
    pub protocol_selection: i32,
    pub upstream_connection_options: Option<UpstreamConnectionOptions>,
    pub close_connections_on_host_health_failure: bool,
    pub ignore_health_on_host_removal: bool,
    pub filters: Vec<Filter>,
    pub load_balancing_policy: Option<LoadBalancingPolicy>,
    pub lrs_server: Option<ConfigSource>,
    pub track_timeout_budgets: bool,
    pub upstream_config: Option<TypedExtensionConfig>,
    pub track_cluster_stats: Option<TrackClusterStats>,
    pub preconnect_policy: Option<PreconnectPolicy>,
    pub connection_pool_per_downstream_connection: bool,
    pub cluster_discovery_type: Option<ClusterDiscoveryType>,
    pub lb_config: Option<LbConfig>,

}

Fields
transport_socket_matches: Vec<TransportSocketMatch>name: Stringalt_stat_name: Stringeds_cluster_config: Option<EdsClusterConfig>connect_timeout: Option<Duration>per_connection_buffer_limit_bytes: Option<u32>lb_policy: i32load_assignment: Option<ClusterLoadAssignment>health_checks: Vec<HealthCheck>max_requests_per_connection: Option<u32>
ðŸ‘ŽDeprecated
circuit_breakers: Option<CircuitBreakers>upstream_http_protocol_options: Option<UpstreamHttpProtocolOptions>
ðŸ‘ŽDeprecated
common_http_protocol_options: Option<HttpProtocolOptions>
ðŸ‘ŽDeprecated
http_protocol_options: Option<Http1ProtocolOptions>
ðŸ‘ŽDeprecated
http2_protocol_options: Option<Http2ProtocolOptions>
ðŸ‘ŽDeprecated
typed_extension_protocol_options: HashMap<String, Any>dns_refresh_rate: Option<Duration>dns_failure_refresh_rate: Option<RefreshRate>respect_dns_ttl: booldns_lookup_family: i32dns_resolvers: Vec<Address>
ðŸ‘ŽDeprecated
use_tcp_for_dns_lookups: bool
ðŸ‘ŽDeprecated
dns_resolution_config: Option<DnsResolutionConfig>
ðŸ‘ŽDeprecated
typed_dns_resolver_config: Option<TypedExtensionConfig>wait_for_warm_on_init: Option<bool>outlier_detection: Option<OutlierDetection>cleanup_interval: Option<Duration>upstream_bind_config: Option<BindConfig>lb_subset_config: Option<LbSubsetConfig>common_lb_config: Option<CommonLbConfig>transport_socket: Option<TransportSocket>metadata: Option<Metadata>protocol_selection: i32
ðŸ‘ŽDeprecated
upstream_connection_options: Option<UpstreamConnectionOptions>close_connections_on_host_health_failure: boolignore_health_on_host_removal: boolfilters: Vec<Filter>load_balancing_policy: Option<LoadBalancingPolicy>lrs_server: Option<ConfigSource>track_timeout_budgets: bool
ðŸ‘ŽDeprecated
upstream_config: Option<TypedExtensionConfig>track_cluster_stats: Option<TrackClusterStats>preconnect_policy: Option<PreconnectPolicy>connection_pool_per_downstream_connection: boolcluster_discovery_type: Option<ClusterDiscoveryType>lb_config: Option<LbConfig>
Implementations

impl Cluster

pub fn lb_policy(&self) -> LbPolicy

Returns the enum value of lb_policy, or the default if the field is set to an invalid enum value.

pub fn set_lb_policy(&mut self, value: LbPolicy)

Sets lb_policy to the provided enum value.

pub fn dns_lookup_family(&self) -> DnsLookupFamily

Returns the enum value of dns_lookup_family, or the default if the field is set to an invalid enum value.

pub fn set_dns_lookup_family(&mut self, value: DnsLookupFamily)

Sets dns_lookup_family to the provided enum value.

pub fn protocol_selection(&self) -> ClusterProtocolSelection

Returns the enum value of protocol_selection, or the default if the field is set to an invalid enum value.

pub fn set_protocol_selection(&mut self, value: ClusterProtocolSelection)

Sets protocol_selection to the provided enum value.