Crate quilkin

Modules

config

Quilkin configuration.

endpoint

Types representing where the data is the sent.

filters

Filters for processing packets.

metadata

xds

Macros

enum_no_match_error

Returns a ConvertProtoConfigError with an error message when an invalid proto enum value was provided in a filter’s proto config.

include_proto

Includes generated Protobuf definitions from tonic.

map_proto_enum

Maps an integer from a protobuf enum value to a target enum variant. Both protobuf and target enum must have similar variants. The protobuf enum variant should be cast-able to an i32 Returns an OK Result with the target enum variant otherwise ConvertProtoConfigError if the provided value does not map to any enum variant.

Structs

Cli

Config

Config is the configuration of a proxy

Proxy

The UDP proxy service.

Type Definitions

Result