Crate quilkin

Re-exports

pub use quilkin_proto as generated;

Modules

cli

codec

Implementations and utility methods for various codecs used in Quilkin.

components

config

Quilkin configuration.

filters

Filters for processing packets.

net

pool

time

xds

Macros

__func_name

Macro that can get the function name of the function the macro is invoked within

include_proto

Includes generated Protobuf definitions from tonic.

temp_file

Creates a temporary file with the specified prefix in a directory named after the calling function, ie using it within a test will place it in a temporary directory named after the test

Structs

Cli

The Command-Line Interface for Quilkin.

Config

Configuration for a component

Proxy

Run Quilkin as a UDP reverse proxy.

Enums

ShutdownKind

Functions

make_shutdown_channel

Type Aliases

Result

ShutdownRx

Receiver for a shutdown event.

ShutdownTx