quilkin::components::proxy::packet_router
Trait PacketMut

pub trait PacketMut: Sized + Packet {
    type FrozenPacket: Packet;

    // Required methods
    fn alloc_sized(&self, size: usize) -> Option<Self>;
    fn as_mut_slice(&mut self) -> &mut [u8] ;
    fn set_len(&mut self, len: usize);
    fn remove_head(&mut self, length: usize);
    fn remove_tail(&mut self, length: usize);
    fn extend_head(&mut self, bytes: &[u8]);
    fn extend_tail(&mut self, bytes: &[u8]);
    fn freeze(self) -> Self::FrozenPacket;
}

Representation of an mutable set of bytes pulled from the network, this trait provides an abstraction over however the packet was received (epoll, io-uring, xdp)

Required Associated Types

type FrozenPacket: Packet

Required Methods

fn alloc_sized(&self, size: usize) -> Option<Self>

fn as_mut_slice(&mut self) -> &mut [u8]

fn set_len(&mut self, len: usize)

fn remove_head(&mut self, length: usize)

fn remove_tail(&mut self, length: usize)

fn extend_head(&mut self, bytes: &[u8])

fn extend_tail(&mut self, bytes: &[u8])

fn freeze(self) -> Self::FrozenPacket

Returns an immutable version of the packet, this allows certain types return a type that can be more cheaply cloned and shared.

Object Safety

This trait is not object safe.