Trait quilkin::filters::capture::CaptureStrategy

pub trait CaptureStrategy {
    fn capture(
        &self, 
        contents: &mut Vec<u8>, 
        metrics: &Metrics
    ) -> Option<Value>;
}

Trait to implement different strategies for capturing packet data.

Required methods

fn capture(&self, contents: &mut Vec<u8>, metrics: &Metrics) -> Option<Value>

Capture packet data from the contents, and optionally returns a value if anything was captured.