quilkin::filters::capture
Trait CaptureStrategy

pub trait CaptureStrategy {
    // Required method
    fn capture(&self, contents: &[u8]) -> Option<(Value, isize)>;
}

Trait to implement different strategies for capturing packet data.

Required Methods

fn capture(&self, contents: &[u8]) -> Option<(Value, isize)>

Capture packet data from the contents, and optionally returns a value if anything was captured.