Trait quilkin::filters::capture::CaptureStrategy

pub trait CaptureStrategy {
    // Required method
    fn capture(&self, contents: &mut PoolBuffer) -> Option<Value>;
}

Trait to implement different strategies for capturing packet data.

Required Methods

fn capture(&self, contents: &mut PoolBuffer) -> Option<Value>

Capture packet data from the contents, and optionally returns a value if anything was captured.