Struct quilkin::filters::FilterSet

pub struct FilterSet(/* private fields */);

A set of filters to be registered with a FilterRegistry.

Current default filters:

debug

local_rate_limit

concatenate

load_balancer

capture

token_router

compress

Implementations

impl FilterSet

pub fn default_with(filters: impl IntoIterator<Item = DynFilterFactory>) -> Self

Returns a FilterSet with the filters provided through filters in addition to the defaults. Any filter factories provided by filters will override any defaults with a matching name.

pub fn with(filters: impl IntoIterator<Item = DynFilterFactory>) -> Self

Creates a new FilterSet with the set of filter_factories without any defaults.

pub fn get(&self, key: &str) -> Option<&Arc<DynFilterFactory>>

Returns a DynFilterFactory if one matches id, otherwise returns None.

pub fn insert( &mut self, value: DynFilterFactory ) -> Option<Arc<DynFilterFactory>>

Inserts factory for the specified FilterFactory, returning any previous filter stored at that location if present.

pub fn iter(&self) -> Iter<'_>

Returns a by reference iterator over the set of filters.