Trait quilkin::filters::FilterFactory

pub trait FilterFactory: Sync + Send {
    fn name(&self) -> &'static str;

    fn config_schema(&self) -> RootSchema;

    fn create_filter(
        &self, 
        args: CreateFilterArgs
    ) -> Result<FilterInstance, Error>;

    fn require_config(
        &self, 
        config: Option<ConfigType>
    ) -> Result<ConfigType, Error> { ... }
}

Provides the name and creation function for a given Filter.

Required methods

fn name(&self) -> &'static str

name returns the configuration name for the Filter The returned string identifies the filter item’s path with the following format: quilkin.filters.<module>.<version>.<item-name> where: : The rust module name containing the filter item : The filter’s version. : The name of the rust item (e.g enum, struct) implementing the filter. For example the v1alpha1 version of the debug filter has the name: quilkin.filters.debug_filter.v1alpha1.Debug

fn config_schema(&self) -> RootSchema

Returns the schema for the configuration of the Filter.

fn create_filter(&self, args: CreateFilterArgs) -> Result<FilterInstance, Error>

Returns a filter based on the provided arguments.

Provided methods

fn require_config(
&self,
config: Option<ConfigType>
) -> Result<ConfigType, Error>

Returns the ConfigType from the provided Option, otherwise it returns Error::MissingConfig if the Option is None.