Module quilkin::net::endpoint

Types representing where the data is the sent.

Re-exports

pub use self::metadata::DynamicMetadata;

Modules

metadata

Structs

Endpoint

A destination endpoint with any associated metadata.

EndpointAddress

A valid socket address. This differs from std::net::SocketAddr, in that it it supports parsing Domain Names in addition to IP addresses. Domain Names are resolved when ToSocketAddrs::to_socket_addrs is called.

Locality

The location of an Endpoint.

Metadata

Metadata specific to endpoints.

Enums

AddressKind

The kind of address, such as Domain Name or IP address. Note that the FromStr implementation doesn’t actually validate that the name is resolvable. Use EndpointAddress for complete address validation.

MetadataError

Type Aliases

EndpointMetadata