quilkin::net::packet

Trait PacketMut

pub trait PacketMut: Sized + Packet {
    type FrozenPacket: Packet;

    // Required methods
    fn remove_head(&mut self, length: usize);
    fn remove_tail(&mut self, length: usize);
    fn extend_head(&mut self, bytes: &[u8]);
    fn extend_tail(&mut self, bytes: &[u8]);
    fn freeze(self) -> Self::FrozenPacket;
}

Representation of an mutable set of bytes pulled from the network, this trait provides an abstraction over however the packet was received (epoll, io-uring, xdp)

Required Associated Types

type FrozenPacket: Packet

Required Methods

fn remove_head(&mut self, length: usize)

fn remove_tail(&mut self, length: usize)

fn extend_head(&mut self, bytes: &[u8])

fn extend_tail(&mut self, bytes: &[u8])

fn freeze(self) -> Self::FrozenPacket

Returns an immutable version of the packet, this allows certain types return a type that can be more cheaply cloned and shared.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.