quilkin::net::sessions

Trait SessionManager

pub trait SessionManager {
    type Packet: Packet;

    // Required method
    fn send(
        &self,
        key: SessionKey,
        contents: &Self::Packet,
    ) -> Result<(), PipelineError>;
}

Responsible for managing sending processed traffic to its destination and tracking metrics and other information about the session.

Required Associated Types

type Packet: Packet

Required Methods

fn send( &self, key: SessionKey, contents: &Self::Packet, ) -> Result<(), PipelineError>