quilkin::net::xdp
Function spawn

pub fn spawn(
    workers: XdpWorkers,
    config: Arc<Config>,
) -> Result<XdpLoop, XdpSpawnError>

The entrypoint into the XDP I/O loop.

This spawns a named thread for each configured XDP socket to run the packet receiving + processing + sending, after which the eBPF program used to route packets to the XDP sockets is attached to the NIC.

Errors

This can fail if threads can not be spawned for some reason (unlikely), the more likely reason for failure is the inability to attach the eBPF program