Type Alias quilkin::net::xds::AdsClient

pub type AdsClient = Client<AggregatedDiscoveryServiceClient<Channel>>;

Aliased Type

struct AdsClient { /* private fields */ }

Implementations

impl AdsClient

pub fn xds_client_stream( &self, config: Arc<Config>, idle_request_interval: Duration ) -> BidirectionalStream<AggregatedDiscoveryServiceClient<Channel>>

Starts a new stream to the xDS management server.

pub async fn delta_subscribe( self, config: Arc<Config>, idle_request_interval: Duration, resources: impl IntoIterator<Item = (ResourceType, Vec<String>)> ) -> Result<DeltaSubscription, Self>

Attempts to start a new delta stream to the xDS management server, if the management server does not support delta xDS we return the client as an error

impl<C: ServiceClient> Client<C>

pub async fn connect( identifier: String, mode: Admin, management_servers: Vec<Endpoint> ) -> Result<Self>